==> ingredientize_asa24_recalls/scripts/02_ingredientize_mixed_foods.ipynb <==
# Import packages
import pandas as pd
import numpy as np

# Load data for foodcodes that were replaced with FNDDS 2018
new_codes = pd.read_csv('fndds_missing_foods_replaced.csv')
# Load ASA24 dataset
asa24 = pd.read_csv('fl100_recalls_qcd.csv')

new_codes['NewFoodCode'].sort_values(ascending=True)

new_codes = new_codes[new_codes['NewFoodCode'] != 9]

asa24 = asa24[asa24['FoodCode'] != 9]
asa24 = asa24[asa24['Food_Description'] != ' ']

asa24['FoodCode'].dropna(inplace=True)

old = new_codes['FoodCode'].tolist()

new = new_codes['NewFoodCode'].tolist()

keep = asa24[~asa24['FoodCode'].isin(old)]

match = asa24[asa24['FoodCode'].isin(old)]

match['FoodCode'].replace(old, new, inplace=True) #SettingwithcopyWarning

asa24_matched = pd.concat([match, keep], axis=0)

### Merge ASA24 updated foodcodes with FNDDS to ingredientize

#Load data
fndds = pd.read_csv('data/fndds/fndds_2018/fndds2018.csv')

#Rename column 'Food code' to 'FoodCode' for merging with ASA data
fndds.columns=['FoodCode',
               'fndds_description',
               'WWEIA Category number',
               'WWEIA Category description',
               'Seq num',
               'Ingredient code',
               'Ingredient description',
               'Ingredient weight (g)',
               'Retention code',
               'Moisture change (%)']

asa_fndds = pd.merge(asa24_matched, fndds, on='FoodCode')

asa_set = set(asa24['Food_Description'].unique())
asa_fndds_set = set(asa_fndds['Food_Description'].unique())

# 2 foods did not map to FNDDS2018
unmapped = asa_set.difference(asa_fndds_set)
unmapped

#Rearrange columns of interest
cols_to_order = ['UserName',
 'FoodCode',
 'Food_Description',
 'fndds_description',
 'WWEIA Category number',
 'WWEIA Category description',
 'FoodAmt',                
 'Seq num',
 'Ingredient code',
 'Ingredient description',
 'Ingredient weight (g)']
new_columns = cols_to_order + (asa_fndds.columns.drop(cols_to_order).tolist())
asa_fndds = asa_fndds[new_columns]

asa_fndds.to_csv('asa24_ingredients/asa_fndds_matched_120721.csv', index=None)
