==> ingredientize_asa24_recalls/scripts/03_ingredient_code_remap.ipynb <==
## The following code is used to deconvolute ingredient codes that are represented by 8-digit food codes for proper ingredientization

import pandas as pd

# Load data
fndds = pd.read_csv('fndds2018.csv')
asa_fndds = pd.read_csv('asa_fndds_matched_120721.csv')

# Extract all ingredient codes represented as 8-digit foodcodes
asa_fndds = asa_fndds[asa_fndds['Ingredient code']>10000000]

asa_fndds.shape

asa_fndds = asa_fndds.drop_duplicates(subset ='Ingredient code')

fndds.drop(columns=['Main food description', 'WWEIA Category number', 'WWEIA Category description', 'Seq num', 'Retention code', 'Moisture change (%)'], inplace=True)

fndds.rename(columns={'Ingredient code':'Ingred_code'}, inplace = True)

fndds.rename(columns={'Food code':'Ingredient code'}, inplace = True)

asa_recode = pd.merge(asa_fndds, fndds, on='Ingredient code')

asa_recode.to_csv('ingred_recode/asa_recode.csv')

asa_code_remap = asa_recode[asa_recode.Ingred_code > 10000000] # get only 8-digit foodcodes

asa_code_remap_unique = asa_code_remap.drop_duplicates(subset = 'Ingred_code')

### Following code is to iterate through the ingredient codes that contain an 8-digit foodcode and merge with the parent foodcode to find embedded ingredient codes, some ingredient codes require mapping iterations

asa_code_remap_unique.to_csv('ingred_recode/unique_codes_to_map_102021.csv')

asa_recode.rename(columns={'Ingred_code':'Ingredient_subcode'}, inplace=True)

fndds.rename(columns={'Ingredient code':'Ingredient_subcode'}, inplace=True)

asa_recode2 = pd.merge(asa_recode, fndds, on='Ingredient_subcode')

asa_recode2.to_csv('ingred_recode/asa_recode2.csv')

asa_recode2.rename(columns={'Ingred_code':'Ingredient_subcode2'}, inplace=True)

fndds.rename(columns={'Ingredient_subcode':'Ingredient_subcode2'}, inplace=True)

asa_recode3 = pd.merge(asa_recode2, fndds, on='Ingredient_subcode2')

asa_recode3.to_csv('ingred_recode/asa_recode3.csv')

asa_recode3.rename(columns={'Ingred_code':'Ingredient_subcode3'}, inplace=True)

fndds.rename(columns={'Ingredient_subcode2':'Ingredient_subcode3'}, inplace=True)

asa_recode4 = pd.merge(asa_recode3, fndds, on='Ingredient_subcode3')

asa_recode4.to_csv('ingred_recode/asa_recode4.csv')

asa_recode4.rename(columns={'Ingred_code':'Ingredient_subcode4'}, inplace=True)

fndds.rename(columns={'Ingredient_subcode3':'Ingredient_subcode4'}, inplace=True)

asa_recode5 = pd.merge(asa_recode4, fndds, on='Ingredient_subcode4')

asa_recode5.to_csv('ingred_recode/asa_recode5.csv')

asa_recode5.rename(columns={'Ingred_code':'Ingredient_subcode5'}, inplace=True)

fndds.rename(columns={'Ingredient_subcode4':'Ingredient_subcode5'}, inplace=True)

asa_recode6 = pd.merge(asa_recode5, fndds, on='Ingredient_subcode5')

asa_recode6.to_csv('ingred_recode/asa_recode6.csv')

# All csv files combined manually to create the final mappings: ingred_code_remapped_102021.csv
