==> ingredientize_asa24_recalls/scripts/04_map_ingredient_nutrient_values.ipynb <==
# Map carbohydrate and fiber content in ASA24 foods from FNDDS 2018

import numpy as np
import pandas as pd

# Load data from FNDDS ingredients containing carobhydrate values
nutrient_values = pd.read_csv('fndds_2018_ingredient_carbohydrate_values.csv')
nutrient_values['Nutrient description'].unique()

## Extract nutrients of interest for ingredients in diet recalls

fiber = nutrient_values[nutrient_values['Nutrient description'] == 'Fiber, total dietary']
fiber = fiber.rename(columns={'Nutrient value':'Fiber (g)'})
fiber = fiber.drop(columns=['Ingredient description', 'Nutrient code', 'Nutrient description'])
fiber = fiber.reset_index()

energy =  nutrient_values[nutrient_values['Nutrient description'] == 'Energy']
energy = energy.rename(columns={'Nutrient value':'Energy (kcal)'})
energy = energy.drop(columns=['Ingredient description', 'Nutrient code', 'Nutrient description'])
energy = energy.reset_index()

carbs = nutrient_values[nutrient_values['Nutrient description'] == 'Carbohydrate']
carbs = carbs.rename(columns={'Nutrient value':'Carbohydrate (g)'})
carbs = carbs.drop(columns=['Ingredient description', 'Nutrient code', 'Nutrient description'])
carbs = carbs.reset_index()

nutrients = pd.concat([fiber, carbs['Carbohydrate (g)'], energy['Energy (kcal)']], axis = 1)

nutrients = nutrients.drop(columns='index')
nutrients

# Load data for ingredientized ASA24
asa24 = pd.read_csv('asa_fndds_matched_120721.csv')

asa24 = asa24[asa24['Food_Description'].notna()]

asa24['uniqueID'] = asa24['UserName'].astype(str) + asa24['FoodCode'].astype(str) + asa24['Ingredient code'].astype(str) + asa24['RecallNo'].astype(str) + asa24['FoodNum'].astype(str) + asa24['CodeNum'].astype(str) + asa24['Occ_No'].astype(str) + asa24['FoodAmt'].astype(str) + asa24['Seq num'].astype(str)

asa24 = asa24[asa24['uniqueID'].notna()]
asa24.shape

# Load remapped food_ingredient code data
code_remap = pd.read_csv('ingred_code_remapped_102021.csv')

# Drop columns to better view data
cols = ['UserName',
        'uniqueID',
 'FoodCode',
 'Food_Description',
 'fndds_description',
 'WWEIA Category number',
 'WWEIA Category description',
 'RecallNo',
 'FoodAmt',                
 'FoodNum',
 'Ingredient code',
 'Ingredient description',
 'Ingredient weight (g)']
asa24 = asa24[cols]

asa_recode = pd.merge(asa24, code_remap, how='outer', on='Ingredient code')

asa_recode = asa_recode[asa_recode['Ingredient code'].isin(code_remap['Ingredient code']) ==True]

asa_recode = asa_recode.drop(columns=['Ingredient code', 'Ingredient description', 'Ingredient weight (g)', 'Ingredient description_x', 'Ingredient weight (g)_x'])

asa_recode = asa_recode.rename(columns={'Ingredient_subcode':'Ingredient code', 'Ingredient description_y': 'Ingredient description', 'Ingredient weight (g)_y': 'Ingredient weight (g)'})

asa24 = asa24[asa24['Ingredient code'] < 10000000]

asa24_all = pd.concat([asa24, asa_recode])

asa24_all.shape

asa_nutrients = pd.merge(asa24_all, nutrients, how='inner', on='Ingredient code')

asa_nutrients.shape

ingred_sum = asa_nutrients.groupby(['UserName', 'FoodCode','uniqueID', 'Food_Description', 'Ingredient code', 'WWEIA Category number', 'WWEIA Category description', 'Ingredient description', 'FoodAmt', 'FoodNum','RecallNo', 'Fiber (g)', 'Carbohydrate (g)', 'Energy (kcal)'], dropna=True)['Ingredient weight (g)'].agg(np.sum)

ingred_df = ingred_sum.reset_index()

ingred_df['Ingredient code'].nunique()

ingred_wts = ingred_sum.groupby(level=['UserName', 'FoodCode', 'FoodNum', 'RecallNo']).transform(np.sum)

ingred_wt = ingred_wts.to_frame()

ingred_wt.rename(columns={'Ingredient weight (g)': 'ingredient_weight_total'}, inplace=True)

ingred_sum = ingred_sum.to_frame()

ingred_wt['Ingredient weight (g)'] = ingred_sum['Ingredient weight (g)']

ingred_wt = ingred_wt.reset_index()

ingred_wt['Proportion of ingredient'] = ingred_wt['Ingredient weight (g)'] / ingred_wt['ingredient_weight_total']

ingred_wt['ingredient_consumed_g'] = ingred_wt['Proportion of ingredient'] * ingred_wt['FoodAmt']

ingred_wt['fiber_consumed_g'] = ingred_wt['ingredient_consumed_g'] * (ingred_wt['Fiber (g)']/100)

ingred_wt['carb_consumed_g'] = ingred_wt['ingredient_consumed_g'] * (ingred_wt['Carbohydrate (g)']/100)

ingred_wt['cal_consumed'] = ingred_wt['ingredient_consumed_g'] * (ingred_wt['Energy (kcal)']/100)

ingred_wt['cal_from_carb'] = ingred_wt['carb_consumed_g']*4

ingred_wt.to_csv('ingredient_fiber_carb_weights_nndc_120721.csv')

# Cells below used for determining the frequency of foods consumed in the cohort. Not used for this project.

ingred_freq = ingred_wt.groupby('Ingredient description')['Ingredient description'].count()

ingred_freq = ingred_freq.sort_values(ascending=False).to_frame()

ingred_freq.rename(columns={'Ingredient description': 'frequency'}).to_csv('ingredient_frequency_nndc_120721.csv')
