==> ingredientize_asa24_recalls/scripts/05_merge_asa24_glycopedia_all_matches.ipynb <==
import pandas as pd
import numpy as np
pd.options.display.max_columns = 39

# Load data
asa24 = pd.read_csv('ingredient_fiber_carb_weights_nndc_120721.csv')
matches = pd.read_csv('asa_glycan_foods_subject_filtered_qcd_122221.csv')
glycopedia = pd.read_csv('glycopedia_wet_wt_040722.csv')
glycopedia = glycopedia[glycopedia['Simple name'] != 'Fuyu persimmon flesh'] # Monosaccharide data not available for persimmon

matches = matches.rename(columns={'Ingredient description_y':'Ingredient description'})
food_list = matches['Ingredient description'].to_list()

matches['Ingredient description'].nunique()

asa24_match = asa24[asa24['Ingredient description'].isin(food_list)]

asa24_200 = pd.merge(asa24_match, matches, on='Ingredient description')

asa24_200.rename(columns={'glycan_food':'Simple name'}, inplace=True)

asa24['Ingredient description'].nunique()

glycopedia = glycopedia.drop(['Food, Varietal, Parts, Form (pear, Bartlet, whole without seeds, fresh) '], axis=1)

asa_glycan = pd.merge(asa24_200, glycopedia, on='Simple name')

cal_per_sub = asa24_200.groupby('UserName')['cal_from_carb'].agg(np.sum) / asa24.groupby('UserName')['cal_from_carb'].agg(np.sum)

cut75 = cal_per_sub.sort_values(ascending=True)[170:]

cut75

cut75 = cut75.index.to_list()

carb_75_percent = asa_glycan[asa_glycan.UserName.isin(cut75)]

glycan_consumed = asa_glycan.loc[:, 'Glucose':'Ribose'].astype(float)

glycan_consumed = glycan_consumed.dropna()

glycan_consumed = glycan_consumed.mul(asa_glycan['ingredient_consumed_g'], axis=0)

asa_glycan = asa_glycan.drop(asa_glycan.loc[:, 'Glucose':'Ribose'], axis=1)

asa_glycan = pd.concat([asa_glycan, glycan_consumed], axis=1)

asa_glycan = asa_glycan.drop(columns='Unnamed: 0')

by_subject_recall = asa_glycan.groupby(['UserName','RecallNo'])[['fiber_consumed_g', 'carb_consumed_g', 'cal_consumed', 'Glucose', 'Galactose', 'Fructose', 'Xylose', 'Arabinose', 'Fucose', 'Rhamnose', 'GlcA', 'GalA', 'GlcNAc', 'GalNAc', 'Mannose', 'Allose', 'Ribose']].agg(np.sum)

average_recall_mean_cal = by_subject_recall.groupby('UserName')[['fiber_consumed_g', 'carb_consumed_g', 'cal_consumed', 'Glucose', 'Galactose', 'Fructose', 'Xylose', 'Arabinose', 'Fucose', 'Rhamnose', 'GlcA', 'GalA', 'GlcNAc', 'GalNAc', 'Mannose', 'Allose', 'Ribose']].agg(np.mean)

average_recall_mean_cal_75 = average_recall_mean_cal.reset_index()

average_recall_mean_cal_75 = average_recall_mean_cal_75[average_recall_mean_cal_75.UserName.isin(cut75)]

average_recall_mean_cal_75.to_csv('asa_glycan_energy_unadjusted_040722.csv', index=None)

average_recall_mean = average_recall_mean_cal.drop(columns=['cal_consumed'])
average_recall_mean = average_recall_mean.iloc[:,0:].div(average_recall_mean_cal['cal_consumed'], axis=0) 
average_recall_mean = average_recall_mean.mul(1000)
average_recall_mean = average_recall_mean.reset_index()
average_recall_mean

average_recall_mean_75 = average_recall_mean[average_recall_mean.UserName.isin(cut75)]

average_recall_mean_cal = average_recall_mean_cal.reset_index()

cal_correct = average_recall_mean_cal['Glucose'] / average_recall_mean['Glucose']

cal_correct.index = average_recall_mean['UserName']

cal_correct

asa_glycan['idx'] = asa_glycan.index

ingred_all = asa_glycan.groupby(['UserName', 'FoodCode','idx', 'Food_Description', 'Ingredient code', 'WWEIA Category number', 'WWEIA Category description', 'Ingredient description', 'FoodAmt', 'FoodNum','RecallNo', 'ingredient_consumed_g', 'Fiber (g)', 'Carbohydrate (g)', 'Energy (kcal)', 'cal_consumed', 'cal_from_carb', 'Simple name', 'glycan_food_class'], dropna=True)[[ 'fiber_consumed_g', 'carb_consumed_g', 'Glucose', 'Galactose', 'Fructose', 'Arabinose', 'Xylose', 'Fucose', 'Rhamnose', 'GlcA', 'GalA', 'GlcNAc', 'GalNAc', 'Mannose', 'Allose', 'Ribose']].agg(np.sum)

unadjusted_intake = ingred_all.reset_index()
unadjusted_intake = unadjusted_intake[unadjusted_intake.UserName.isin(cut75)]

unadjusted_intake['Simple name'].nunique()

unadjusted_intake['Ingredient code'].nunique()

supp_table_1 = unadjusted_intake.drop_duplicates(subset = ['Ingredient description'])

supp_table_1.to_csv('supp_table_1.csv')

unadjusted_intake.to_csv('all_items_unadjusted_041122.csv', index=None)

ingred_all = ingred_all.div(cal_correct, axis = 0, level = 0)

ingred_all = ingred_all.reset_index()
ingred_all = ingred_all[ingred_all.UserName.isin(cut75)]

ingred_all = ingred_all.reset_index()
ingred_all = ingred_all[ingred_all.UserName.isin(cut75)]
ingred_all = ingred_all.drop(columns=['index'])
#ingred_all = pd.merge(ingred_all, wweia, on = 'WWEIA Category number')

#ingred_all= ingred_all.drop(columns=['level_0'])

ingred_all.to_csv('all_items_cal_adjusted_041122.csv', index=None)
